Optimize count retrieval, fix warnings

+++ DX12Library/src/Camera.cpp
@@ -162,7 +162,7 @@ Camera::Frustum Camera::GetFrustum() const
     frustum.m_Planes[Frustum::NearPlane] = FrustumPlane(XMVectorSet(0, 0, 1, 0), XMVectorSet(0, 0, ZNear, 1), toWorld);
     frustum.m_Planes[Frustum::FarPlane] = FrustumPlane(XMVectorSet(0, 0, -1, 0), XMVectorSet(0, 0, ZFar, 1), toWorld);
 
-    const float hh = tan(VFov * 0.5f);
+    const float hh = tanf(VFov * 0.5f);
     const float hw = hh * AspectRatio;
 
     const auto nw = XMVector3Normalize(XMVectorSet(-hw, hh, 1, 0));
